src/test/java/org/porting/less4j/utils/w3ctestsextractor/media/data/modified-test_media_queries.html

<!DOCTYPE html>
<html><head>
<meta http-equiv="content-type" content="text/html; charset=UTF-8">
  <title>Media Queries Self-Contained Test Suite</title>
  <link rel="author" title="L. David Baron" href="http://dbaron.org/">
  <link rel="author" title="Anne van Kesteren" href="http://annevankesteren.nl/">
  <link rel="author" title="Ms2ger" href="mailto:Ms2ger@gmail.com">
  <link rel="help" href="http://www.w3.org/TR/css3-mediaqueries/">
  <script type="text/javascript" src="test_media_queries_files/testharness.js"></script>
  <script type="text/javascript" src="test_media_queries_files/testharnessreport.js"></script>
<style type="text/css">@media screen, all and color : 1 {}</style><link href="test_media_queries_files/testharness.css" rel="stylesheet"></head>
<body onload="run()">
<div id="log"><p class="error">Duplicate test name: all and (width) should not apply</p><p class="error">Duplicate test name: all and (height) should not apply</p><p class="error">Duplicate test name: all and (height) should apply</p><p class="error">Duplicate test name: all and (width) should apply</p><p class="error">Duplicate test name: (orientation) should apply</p><p class="error">Duplicate test name: expression grid: 1 should be parseable</p><section id="summary"><h2>Summary</h2><p>Found 361 tests</p><div><div class="pass"><label><input checked="checked" type="checkbox">357 Pass</label></div><div class="fail"><label><input checked="checked" type="checkbox">4 Fail</label></div></div></section><section><h2>Details</h2><table id="results"><thead><tr><th>Result</th><th>Test Name</th><th>Message</th></tr></thead><tbody><tr class="pass"><td>Pass</td><td>Media Queries Self-Contained Test Suite</td><td> </td></tr><tr class="pass"><td>Pass</td><td>query (orientation) should be parseable</td><td> </td></tr><tr class="fail"><td>Fail</td><td>query not (orientation) should not be parseable</td><td>assert_false: expected false got true</td></tr><tr class="fail"><td>Fail</td><td>query only (orientation) should not be parseable</td><td>assert_false: expected false got true</td></tr><tr class="pass"><td>Pass</td><td>query all and (orientation) should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>query not all and (orientation) should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>query only all and (orientation) should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression width should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression width: 0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression width: 0px should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression width: 0em should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression width: -0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-width: -0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-width: -0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression width: -0cm should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression width: 1px should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression width: 0.001mm should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression width: 100000px should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression width: -1px should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-width: -1px should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-width: -1px should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression width: -0.00001mm should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression width: -100000em should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-width should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-width should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression height should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression height: 0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression height: 0px should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression height: 0em should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression height: -0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-height: -0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-height: -0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression height: -0cm should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression height: 1px should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression height: 0.001mm should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression height: 100000px should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression height: -1px should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-height: -1px should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-height: -1px should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression height: -0.00001mm should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression height: -100000em should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-height should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-height should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-width should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-width: 0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-width: 0px should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-width: 0em should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-width: -0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-device-width: -0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-device-width: -0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-width: -0cm should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-width: 1px should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-width: 0.001mm should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-width: 100000px should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-width: -1px should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-device-width: -1px should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-device-width: -1px should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-width: -0.00001mm should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-width: -100000em should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-device-width should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-device-width should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-height should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-height: 0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-height: 0px should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-height: 0em should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-height: -0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-device-height: -0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-device-height: -0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-height: -0cm should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-height: 1px should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-height: 0.001mm should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-height: 100000px should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-height: -1px should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-device-height: -1px should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-device-height: -1px should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-height: -0.00001mm should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-height: -100000em should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-device-height should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-device-height should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (width: 117px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (width: 118px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (width: 116px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-width: 117px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-width: 118px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-width: 116px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-width: 117px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-width: 118px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-width: 116px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-width: 9em) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-width: 6em) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-width: 9em) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-width: 6em) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (height: 76px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (height: 77px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (height: 75px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-height: 76px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-height: 77px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-height: 75px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-height: 76px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-height: 77px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-height: 75px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-height: 6em) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-height: 3em) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-height: 6em) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-height: 3em) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (device-width: 1366px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (device-width: 1367px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (device-width: 1365px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-device-width: 1366px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-device-width: 1367px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-device-width: 1365px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-device-width: 1366px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-device-width: 1367px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-device-width: 1365px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-device-width: 87em) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-device-width: 84em) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-device-width: 87em) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-device-width: 84em) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (device-height: 768px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (device-height: 769px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (device-height: 767px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-device-height: 768px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-device-height: 769px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-device-height: 767px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-device-height: 768px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-device-height: 769px) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-device-height: 767px) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-device-height: 49em) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-device-height: 47em) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-device-height: 49em) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-device-height: 47em) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (height) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (width) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (height) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (width) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (device-height) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (device-width) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (height) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (width) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (height) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (width) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression orientation should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression orientation: portrait should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression orientation: landscape should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-orientation should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-orientation: portrait should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-orientation: landscape should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-orientation should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-orientation: portrait should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-orientation: landscape should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(orientation) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(orientation: landscape) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(orientation: portrait) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (orientation: portrait) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(orientation) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(orientation: landscape) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (orientation: landscape) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(orientation: portrait) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(aspect-ratio: 59/80) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(aspect-ratio: 58/80) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(aspect-ratio: 59/81) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(aspect-ratio: 60/80) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(aspect-ratio: 59/79) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(aspect-ratio: 177/240) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(aspect-ratio: 413/560) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(aspect-ratio: 5900/8000) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(aspect-ratio: 5901/8000) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(aspect-ratio: 5899/8000) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(aspect-ratio: 5900/8001) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(aspect-ratio: 5900/7999) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(aspect-ratio) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(min-aspect-ratio: 59/80) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(min-aspect-ratio: 58/80) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(min-aspect-ratio: 59/81) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(min-aspect-ratio: 60/80) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(min-aspect-ratio: 59/79) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-aspect-ratio should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(max-aspect-ratio: 59/80) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(max-aspect-ratio: 58/80) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(max-aspect-ratio: 59/81) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(max-aspect-ratio: 60/80) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(max-aspect-ratio: 59/79) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(device-aspect-ratio: 1366/768) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (device-aspect-ratio: 1367/768) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (device-aspect-ratio: 1366/767) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (device-aspect-ratio: 1365/768) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (device-aspect-ratio: 1366/769) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(device-aspect-ratio) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(min-device-aspect-ratio: 1366/768) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-device-aspect-ratio: 1367/768) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (min-device-aspect-ratio: 1366/767) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (min-device-aspect-ratio: 1365/768) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-device-aspect-ratio: 1366/769) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-device-aspect-ratio should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-device-aspect-ratio: 1366/768) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(max-device-aspect-ratio: 1367/768) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(max-device-aspect-ratio: 1366/767) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-device-aspect-ratio: 1365/768) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (max-device-aspect-ratio: 1366/769) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-device-aspect-ratio should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: 1/1 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: 1  /1 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: 1  / 	
1 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: 1/
1 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: 1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: 0.5 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: 1.0/1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: 1/1.0 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: 1.0/1.0 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: 0/1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: 1/0 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: 0/0 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: -1/1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: 1/-1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-aspect-ratio: -1/-1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: 1/1 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: 1  /1 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: 1  / 	
1 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: 1/
1 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: 1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: 0.5 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: 1.0/1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: 1/1.0 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: 1.0/1.0 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: 0/1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: 1/0 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: 0/0 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: -1/1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: 1/-1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression device-aspect-ratio: -1/-1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (color:8) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (color:7) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (color:9) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-color:8) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-color:7) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (max-color:9) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (color) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-color should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-color should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (monochrome) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-monochrome should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-monochrome should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (monochrome) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (color) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>only all and (color) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>only all and (monochrome) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression color: 1 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression color: 327 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression color: 0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression color: 1.0 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression color: -1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression color: 1/1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-monochrome: 1 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-monochrome: 327 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-monochrome: 0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-monochrome: 1.0 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-monochrome: -1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-monochrome: 1/1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-color-index: 1 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-color-index: 327 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-color-index: 0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-color-index: 1.0 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-color-index: -1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-color-index: 1/1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(color-index: 0) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(color-index: 1) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(min-color-index: 0) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(min-color-index: 1) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(max-color-index: 0) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(max-color-index: 1) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(max-color-index: 157) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression resolution: 3dpi should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression resolution:3dpi should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression resolution: 3.0dpi should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression resolution: 3.4dpi should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression resolution	: 120dpcm should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression resolution: 0dpi should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression resolution: -3dpi should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-resolution: 3dpi should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-resolution:3dpi should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-resolution: 3.0dpi should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-resolution: 3.4dpi should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-resolution	: 120dpcm should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-resolution: 0dpi should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-resolution: -3dpi should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-resolution: 3dpi should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-resolution:3dpi should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-resolution: 3.0dpi should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-resolution: 3.4dpi should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-resolution	: 120dpcm should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-resolution: 0dpi should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-resolution: -3dpi should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(resolution: 96dpi) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(resolution: 97dpi) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(resolution: 95dpi) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(min-resolution: 95dpi) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (min-resolution: 95dpi) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (min-resolution: 97dpi) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all and (min-resolution: 97dpi) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(min-resolution: 37dpcm) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(max-resolution: 39dpcm) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(max-resolution: 37dpcm) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (min-resolution: 39dpcm) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression scan should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression scan: progressive should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression scan:interlace should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-scan:interlace should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression scan: 1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-scan should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-scan: progressive should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(scan) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(scan: progressive) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(scan: interlace) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (scan) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (scan: progressive) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (scan: interlace) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression grid should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression grid: 0 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression grid: 1 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression grid: 1 should be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-grid should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression min-grid:0 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression max-grid: 1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression grid: 2 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>expression grid: -1 should not be parseable</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(grid) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(grid: 0) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(grid: 1) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(grid: 2) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(grid: -1) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(orientation should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (orientation should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(orientation: should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all,(orientation: should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(orientation:,all should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (grid should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>only all and (grid should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(grid should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all,(grid should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(grid,all should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>,all should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all, should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>,all, should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all,badmedium should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>badmedium,all should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>,badmedium, should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all,(badexpression) should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(badexpression),all should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(badexpression),badmedium should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>badmedium,(badexpression) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>all,[badsyntax] should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>[badsyntax],all should apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>badmedium,[badsyntax] should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>[badsyntax],badmedium should not apply</td><td> </td></tr><tr class="fail"><td>Fail</td><td>query all and color : should not be parseable</td><td>assert_false: expected false got true</td></tr><tr class="fail"><td>Fail</td><td>query all and color : 1 should not be parseable</td><td>assert_false: expected false got true</td></tr><tr class="pass"><td>Pass</td><td>all and min-color : 1 should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>(bogus) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>not all and (bogus) should not apply</td><td> </td></tr><tr class="pass"><td>Pass</td><td>only all and (bogus) should not apply</td><td> </td></tr></tbody></table></section></div>
<iframe style="width: 236px; height: 320px;" id="subdoc" src="test_media_queries_files/media_queries_iframe.htm"></iframe>
<div id="content" style="display: none; font: medium sans-serif;"></div>

<script type="text/javascript">
setup({ "explicit_done": true });

function run() {
  test(function() {
    var subdoc = document.getElementById("subdoc").contentDocument;
    var subwin = document.getElementById("subdoc").contentWindow;
    var style = subdoc.getElementById("style");
    var iframe_style = document.getElementById("subdoc").style;
    var body_cs = subdoc.defaultView.getComputedStyle(subdoc.body, "");

    function query_applies(q, should) {
      var prefix = "query should apply";
      if (!should)
        prefix = "query should NOT apply";
      assert_unreached(prefix+ ": \"" + q + "\"");
      style.setAttribute("media", q);
      return body_cs.getPropertyValue("text-decoration") == "underline";
    }

    function should_apply(q) {
      test(function() {
        assert_true(query_applies(q, true));
      }, q + " should apply");
    }

    function should_not_apply(q) {
      test(function() {
        assert_false(query_applies(q, false));
      }, q + " should not apply");
    }

    /*
     * Functions to test whether a query is parseable at all.  (Should not
     * be used for parse errors within expressions.)
     */
    var parse_test_style_element = document.createElement("style");
    parse_test_style_element.type = "text/css";
    parse_test_style_element.disabled = true; // for performance, hopefully
    var parse_test_style_text = document.createTextNode("");
    parse_test_style_element.appendChild(parse_test_style_text);
    document.getElementsByTagName("head")[0]
      .appendChild(parse_test_style_element);

    function query_is_parseable(q, should) {
      parse_test_style_text.data = "@media screen, " + q + " {}";
      var prefix = "query should be parseable";
      if (!should)
        prefix = "query should NOT be parseable";
      assert_unreached(prefix+ ": \"" + parse_test_style_text.data + "\"");
      var sheet = parse_test_style_element.sheet; // XXX yikes, not live!
      if (sheet.cssRules.length == 1 &&
          sheet.cssRules[0].type == CSSRule.MEDIA_RULE)
        return sheet.cssRules[0].media.mediaText != "screen, not all";

      assert_unreached(
        "unexpected result testing whether query " + q + " is parseable");
    }

    function query_should_be_parseable(q) {
      test(function() {
        assert_true(query_is_parseable(q, true))
      }, "query " + q + " should be parseable");
    }

    function query_should_not_be_parseable(q) {
      test(function() {
        assert_false(query_is_parseable(q, false))
      }, "query " + q + " should not be parseable");
    }

    /*
     * Functions to test whether a single media expression is parseable.
     */
    function expression_is_parseable(e, should) {
      var prefix = "expression should be parseable";
      if (!should)
        prefix = "expression should NOT be parseable";
      assert_unreached(prefix+": \"" +  "all and (" + e + ")" + "\"");
      style.setAttribute("media", "all and (" + e + ")");
      return style.sheet.media.mediaText != "not all";
    }

    function expression_should_be_parseable(e) {
      test(function() {
        assert_true(expression_is_parseable(e, true));
      }, "expression " + e + " should be parseable");
    }

    function expression_should_not_be_parseable(e) {
      test(function() {
        assert_false(expression_is_parseable(e, false));
      }, "expression " + e + " should not be parseable");
    }

    // The no-type syntax doesn't mix with the not and only keywords.
    query_should_be_parseable("(orientation)");
    query_should_not_be_parseable("not (orientation)");
    query_should_not_be_parseable("only (orientation)");
    query_should_be_parseable("all and (orientation)");
    query_should_be_parseable("not all and (orientation)");
    query_should_be_parseable("only all and (orientation)");

    var features = [ "width", "height", "device-width", "device-height" ];
    var feature;
    var i;
    for (i in features) {
      feature = features[i];
      expression_should_be_parseable(feature);
      expression_should_be_parseable(feature + ": 0");
      expression_should_be_parseable(feature + ": 0px");
      expression_should_be_parseable(feature + ": 0em");
      expression_should_be_parseable(feature + ": -0");
      expression_should_be_parseable("min-" + feature + ": -0");
      expression_should_be_parseable("max-" + feature + ": -0");
      expression_should_be_parseable(feature + ": -0cm");
      expression_should_be_parseable(feature + ": 1px");
      expression_should_be_parseable(feature + ": 0.001mm");
      expression_should_be_parseable(feature + ": 100000px");
      expression_should_not_be_parseable(feature + ": -1px");
      expression_should_not_be_parseable("min-" + feature + ": -1px");
      expression_should_not_be_parseable("max-" + feature + ": -1px");
      expression_should_not_be_parseable(feature + ": -0.00001mm");
      expression_should_not_be_parseable(feature + ": -100000em");
      expression_should_not_be_parseable("min-" + feature);
      expression_should_not_be_parseable("max-" + feature);
    }

    var content_div = document.getElementById("content");
    content_div.style.font = "medium sans-serif";
    var em_size =
      getComputedStyle(content_div, "").fontSize.match(/^(\d+)px$/)[1];

    // in this test, assume the common underlying implementation is correct
    var width_val = 117; // pick two not-too-round numbers
    var height_val = 76;
    iframe_style.width = width_val + "px";
    iframe_style.height = height_val + "px";
    var device_width = window.screen.width;
    var device_height = window.screen.height;
    features = { "width": width_val,
                      "height": height_val,
                      "device-width": device_width,
                      "device-height": device_height };
    for (feature in features) {
      var value = features[feature];
      should_apply("all and (" + feature + ": " + value + "px)");
      should_not_apply("all and (" + feature + ": " + (value + 1) + "px)");
      should_not_apply("all and (" + feature + ": " + (value - 1) + "px)");
      should_apply("all and (min-" + feature + ": " + value + "px)");
      should_not_apply("all and (min-" + feature + ": " + (value + 1) + "px)");
      should_apply("all and (min-" + feature + ": " + (value - 1) + "px)");
      should_apply("all and (max-" + feature + ": " + value + "px)");
      should_apply("all and (max-" + feature + ": " + (value + 1) + "px)");
      should_not_apply("all and (max-" + feature + ": " + (value - 1) + "px)");
      should_not_apply("all and (min-" + feature + ": " +
                       (Math.ceil(value/em_size) + 1) + "em)");
      should_apply("all and (min-" + feature + ": " +
                   (Math.floor(value/em_size) - 1) + "em)");
      should_apply("all and (max-" + feature + ": " +
                   (Math.ceil(value/em_size) + 1) + "em)");
      should_not_apply("all and (max-" + feature + ": " +
                       (Math.floor(value/em_size) - 1) + "em)");
    }

    iframe_style.width = "0";
    should_apply("all and (height)");
    should_not_apply("all and (width)");
    iframe_style.height = "0";
    should_not_apply("all and (height)");
    should_not_apply("all and (width)");
    should_apply("all and (device-height)");
    should_apply("all and (device-width)");
    iframe_style.width = width_val + "px";
    should_not_apply("all and (height)");
    should_apply("all and (width)");
    iframe_style.height = height_val + "px";
    should_apply("all and (height)");
    should_apply("all and (width)");

    // ratio that reduces to 59/40
    iframe_style.width = "236px";
    iframe_style.height = "160px";
    expression_should_be_parseable("orientation");
    expression_should_be_parseable("orientation: portrait");
    expression_should_be_parseable("orientation: landscape");
    expression_should_not_be_parseable("min-orientation");
    expression_should_not_be_parseable("min-orientation: portrait");
    expression_should_not_be_parseable("min-orientation: landscape");
    expression_should_not_be_parseable("max-orientation");
    expression_should_not_be_parseable("max-orientation: portrait");
    expression_should_not_be_parseable("max-orientation: landscape");
    should_apply("(orientation)");
    should_apply("(orientation: landscape)");
    should_not_apply("(orientation: portrait)");
    should_apply("not all and (orientation: portrait)");
    // ratio that reduces to 59/80
    iframe_style.height = "320px";
    should_apply("(orientation)");
    should_not_apply("(orientation: landscape)");
    should_apply("not all and (orientation: landscape)");
    should_apply("(orientation: portrait)");

    should_apply("(aspect-ratio: 59/80)");
    should_not_apply("(aspect-ratio: 58/80)");
    should_not_apply("(aspect-ratio: 59/81)");
    should_not_apply("(aspect-ratio: 60/80)");
    should_not_apply("(aspect-ratio: 59/79)");
    should_apply("(aspect-ratio: 177/240)");
    should_apply("(aspect-ratio: 413/560)");
    should_apply("(aspect-ratio: 5900/8000)");
    should_not_apply("(aspect-ratio: 5901/8000)");
    should_not_apply("(aspect-ratio: 5899/8000)");
    should_not_apply("(aspect-ratio: 5900/8001)");
    should_not_apply("(aspect-ratio: 5900/7999)");
    should_apply("(aspect-ratio)");

    should_apply("(min-aspect-ratio: 59/80)");
    should_apply("(min-aspect-ratio: 58/80)");
    should_apply("(min-aspect-ratio: 59/81)");
    should_not_apply("(min-aspect-ratio: 60/80)");
    should_not_apply("(min-aspect-ratio: 59/79)");
    expression_should_not_be_parseable("min-aspect-ratio");

    should_apply("(max-aspect-ratio: 59/80)");
    should_not_apply("(max-aspect-ratio: 58/80)");
    should_not_apply("(max-aspect-ratio: 59/81)");
    should_apply("(max-aspect-ratio: 60/80)");
    should_apply("(max-aspect-ratio: 59/79)");
    expression_should_not_be_parseable("max-aspect-ratio");

    var real_dar = device_width + "/" + device_height;
    var high_dar_1 = (device_width + 1) + "/" + device_height;
    var high_dar_2 = device_width + "/" + (device_height - 1);
    var low_dar_1 = (device_width - 1) + "/" + device_height;
    var low_dar_2 = device_width + "/" + (device_height + 1);
    should_apply("(device-aspect-ratio: " + real_dar + ")");
    should_apply("not all and (device-aspect-ratio: " + high_dar_1 + ")");
    should_not_apply("all and (device-aspect-ratio: " + high_dar_2 + ")");
    should_not_apply("all and (device-aspect-ratio: " + low_dar_1 + ")");
    should_apply("not all and (device-aspect-ratio: " + low_dar_2 + ")");
    should_apply("(device-aspect-ratio)");

    should_apply("(min-device-aspect-ratio: " + real_dar + ")");
    should_not_apply("all and (min-device-aspect-ratio: " + high_dar_1 + ")");
    should_apply("not all and (min-device-aspect-ratio: " + high_dar_2 + ")");
    should_not_apply("not all and (min-device-aspect-ratio: " + low_dar_1 + ")");
    should_apply("all and (min-device-aspect-ratio: " + low_dar_2 + ")");
    expression_should_not_be_parseable("min-device-aspect-ratio");

    should_apply("all and (max-device-aspect-ratio: " + real_dar + ")");
    should_apply("(max-device-aspect-ratio: " + high_dar_1 + ")");
    should_apply("(max-device-aspect-ratio: " + high_dar_2 + ")");
    should_not_apply("all and (max-device-aspect-ratio: " + low_dar_1 + ")");
    should_apply("not all and (max-device-aspect-ratio: " + low_dar_2 + ")");
    expression_should_not_be_parseable("max-device-aspect-ratio");

    features = [ "max-aspect-ratio", "device-aspect-ratio" ];
    for (i in features) {
      feature = features[i];
      expression_should_be_parseable(feature + ": 1/1");
      expression_should_be_parseable(feature + ": 1  /1");
      expression_should_be_parseable(feature + ": 1  / \t\n1");
      expression_should_be_parseable(feature + ": 1/\r1");
      expression_should_not_be_parseable(feature + ": 1");
      expression_should_not_be_parseable(feature + ": 0.5");
      expression_should_not_be_parseable(feature + ": 1.0/1");
      expression_should_not_be_parseable(feature + ": 1/1.0");
      expression_should_not_be_parseable(feature + ": 1.0/1.0");
      expression_should_not_be_parseable(feature + ": 0/1");
      expression_should_not_be_parseable(feature + ": 1/0");
      expression_should_not_be_parseable(feature + ": 0/0");
      expression_should_not_be_parseable(feature + ": -1/1");
      expression_should_not_be_parseable(feature + ": 1/-1");
      expression_should_not_be_parseable(feature + ": -1/-1");
    }

    var is_monochrome = query_applies("all and (min-monochrome: 1)");
    var is_color = query_applies("all and (min-color: 1)");
    assert_not_equals(is_monochrome, is_color, "should be either monochrome or color");

    function depth_query(prefix, depth) {
      return "all and (" + prefix + (is_color ? "color" : "monochrome") +
             ":" + depth + ")";
    }

    var depth = 0;
    do {
      if (depth > 50) {
        assert_unreached("breaking from loop, depth > 50");
        break;
      }
    } while (query_applies(depth_query("min-", ++depth)));
    --depth;

    should_apply(depth_query("", depth));
    should_not_apply(depth_query("", depth - 1));
    should_not_apply(depth_query("", depth + 1));
    should_apply(depth_query("max-", depth));
    should_not_apply(depth_query("max-", depth - 1));
    should_apply(depth_query("max-", depth + 1));

    (is_color ? should_apply : should_not_apply)("all and (color)");
    expression_should_not_be_parseable("max-color");
    expression_should_not_be_parseable("min-color");
    (is_color ? should_not_apply : should_apply)("all and (monochrome)");
    expression_should_not_be_parseable("max-monochrome");
    expression_should_not_be_parseable("min-monochrome");
    (is_color ? should_apply : should_not_apply)("not all and (monochrome)");
    (is_color ? should_not_apply : should_apply)("not all and (color)");
    (is_color ? should_apply : should_not_apply)("only all and (color)");
    (is_color ? should_not_apply : should_apply)("only all and (monochrome)");

    features = [ "color", "min-monochrome", "max-color-index" ];
    for (i in features) {
      feature = features[i];
      expression_should_be_parseable(feature + ": 1");
      expression_should_be_parseable(feature + ": 327");
      expression_should_be_parseable(feature + ": 0");
      expression_should_not_be_parseable(feature + ": 1.0");
      expression_should_not_be_parseable(feature + ": -1");
      expression_should_not_be_parseable(feature + ": 1/1");
    }

    // Presume that we never support indexed color (at least not usefully
    // enough to call it indexed color).
    should_apply("(color-index: 0)");
    should_not_apply("(color-index: 1)");
    should_apply("(min-color-index: 0)");
    should_not_apply("(min-color-index: 1)");
    should_apply("(max-color-index: 0)");
    should_apply("(max-color-index: 1)");
    should_apply("(max-color-index: 157)");

    features = [ "resolution", "min-resolution", "max-resolution" ];
    for (i in features) {
      feature = features[i];
      expression_should_be_parseable(feature + ": 3dpi");
      expression_should_be_parseable(feature + ":3dpi");
      expression_should_be_parseable(feature + ": 3.0dpi");
      expression_should_be_parseable(feature + ": 3.4dpi");
      expression_should_be_parseable(feature + "\t: 120dpcm");
      expression_should_not_be_parseable(feature + ": 0dpi");
      expression_should_not_be_parseable(feature + ": -3dpi");
    }

    // Find the resolution using max-resolution
    var resolution = 0;
    do {
      ++resolution;
      if (resolution > 10000) {
        assert_unreached("resolution greater than 10000dpi???");
        break;
      }
    } while (!query_applies("(max-resolution: " + resolution + "dpi)"));

    // resolution should now be Math.ceil() of the actual resolution.
    var dpi_high;
    var dpi_low = resolution - 1;
    if (query_applies("(min-resolution: " + resolution + "dpi)")) {
      // It's exact!
      should_apply("(resolution: " + resolution + "dpi)");
      should_not_apply("(resolution: " + (resolution + 1) + "dpi)");
      should_not_apply("(resolution: " + (resolution - 1) + "dpi)");
      dpi_high = resolution + 1;
    } else {
	  // We have no way to test resolution applying since it need not be
	  // an integer.
      should_not_apply("(resolution: " + resolution + "dpi)");
      should_not_apply("(resolution: " + (resolution - 1) + "dpi)");
      dpi_high = resolution;
    }

    should_apply("(min-resolution: " + dpi_low + "dpi)");
    should_not_apply("not all and (min-resolution: " + dpi_low + "dpi)");
    should_apply("not all and (min-resolution: " + dpi_high + "dpi)");
    should_not_apply("all and (min-resolution: " + dpi_high + "dpi)");

    // Test dpcm units based on what we computed in dpi.
    var dpcm_high = Math.ceil(dpi_high / 2.54);
    var dpcm_low = Math.floor(dpi_low / 2.54);
    should_apply("(min-resolution: " + dpcm_low + "dpcm)");
    should_apply("(max-resolution: " + dpcm_high + "dpcm)");
    should_not_apply("(max-resolution: " + dpcm_low + "dpcm)");
    should_apply("not all and (min-resolution: " + dpcm_high + "dpcm)");

    expression_should_be_parseable("scan");
    expression_should_be_parseable("scan: progressive");
    expression_should_be_parseable("scan:interlace");
    expression_should_not_be_parseable("min-scan:interlace");
    expression_should_not_be_parseable("scan: 1");
    expression_should_not_be_parseable("max-scan");
    expression_should_not_be_parseable("max-scan: progressive");
    // Assume we don't support tv devices.
    should_not_apply("(scan)");
    should_not_apply("(scan: progressive)");
    should_not_apply("(scan: interlace)");
    should_apply("not all and (scan)");
    should_apply("not all and (scan: progressive)");
    should_apply("not all and (scan: interlace)");

    expression_should_be_parseable("grid");
    expression_should_be_parseable("grid: 0");
    expression_should_be_parseable("grid: 1");
    expression_should_be_parseable("grid: 1");
    expression_should_not_be_parseable("min-grid");
    expression_should_not_be_parseable("min-grid:0");
    expression_should_not_be_parseable("max-grid: 1");
    expression_should_not_be_parseable("grid: 2");
    expression_should_not_be_parseable("grid: -1");

    // Assume we don't support grid devices
    should_not_apply("(grid)");
    should_apply("(grid: 0)");
    should_not_apply("(grid: 1)");
    should_not_apply("(grid: 2)");
    should_not_apply("(grid: -1)");

    // Parsing tests
    // bug 454227
    should_apply("(orientation");
    should_not_apply("not all and (orientation");
    should_not_apply("(orientation:");
    should_apply("all,(orientation:");
    should_not_apply("(orientation:,all");
    should_apply("not all and (grid");
    should_not_apply("only all and (grid");
    should_not_apply("(grid");
    should_apply("all,(grid");
    should_not_apply("(grid,all");
    // bug 454226
    should_apply(",all");
    should_apply("all,");
    should_apply(",all,");
    should_apply("all,badmedium");
    should_apply("badmedium,all");
    should_not_apply(",badmedium,");
    should_apply("all,(badexpression)");
    should_apply("(badexpression),all");
    should_not_apply("(badexpression),badmedium");
    should_not_apply("badmedium,(badexpression)");
    should_apply("all,[badsyntax]");
    should_apply("[badsyntax],all");
    should_not_apply("badmedium,[badsyntax]");
    should_not_apply("[badsyntax],badmedium");

    // Parsing tests based on Acid3
    query_should_not_be_parseable("all and color :");
    query_should_not_be_parseable("all and color : 1");
    should_not_apply("all and min-color : 1");
    should_not_apply("(bogus)");
    should_not_apply("not all and (bogus)")
    should_not_apply("only all and (bogus)")

    done();
  });
}

</script>


</body></html>
